backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via
    #   -r requirements.in
    #   django
black==23.3.0
    # via -r requirements.in
build==0.10.0
    # via
    #   -r requirements.in
    #   pip-tools
cfgv==3.3.1
    # via
    #   -r requirements.in
    #   pre-commit
click==8.1.3
    # via
    #   -r requirements.in
    #   black
    #   pip-tools
distlib==0.3.6
    # via
    #   -r requirements.in
    #   virtualenv
django==4.2.1
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
django-cors-headers==4.0.0
    # via -r requirements.in
django-filter==23.2
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.2.2
    # via -r requirements.in
filelock==3.12.0
    # via
    #   -r requirements.in
    #   virtualenv
flake8==6.0.0
    # via
    #   -r requirements.in
    #   flake8-broken-line
    #   flake8-isort
    #   pep8-naming
flake8-broken-line==1.0.0
    # via -r requirements.in
flake8-isort==6.0.0
    # via -r requirements.in
flake8-plugin-utils==1.3.2
    # via
    #   -r requirements.in
    #   flake8-return
flake8-return==1.2.0
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
identify==2.5.24
    # via
    #   -r requirements.in
    #   pre-commit
isort==5.12.0
    # via
    #   -r requirements.in
    #   flake8-isort
mccabe==0.7.0
    # via
    #   -r requirements.in
    #   flake8
mypy-extensions==1.0.0
    # via
    #   -r requirements.in
    #   black
nodeenv==1.8.0
    # via
    #   -r requirements.in
    #   pre-commit
packaging==23.1
    # via
    #   -r requirements.in
    #   black
    #   build
pathspec==0.11.1
    # via
    #   -r requirements.in
    #   black
pep8-naming==0.13.3
    # via -r requirements.in
pillow==9.5.0
    # via -r requirements.in
pip-tools==6.13.0
    # via -r requirements.in
platformdirs==3.5.1
    # via
    #   -r requirements.in
    #   black
    #   virtualenv
pre-commit==3.3.2
    # via -r requirements.in
psycopg2-binary==2.9.6
    # via -r requirements.in
pycodestyle==2.10.0
    # via
    #   -r requirements.in
    #   flake8
pyflakes==3.0.1
    # via
    #   -r requirements.in
    #   flake8
pyjwt==2.7.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
pyproject-hooks==1.0.0
    # via
    #   -r requirements.in
    #   build
pytz==2023.3
    # via
    #   -r requirements.in
    #   djangorestframework
pyyaml==6.0
    # via
    #   -r requirements.in
    #   pre-commit
sqlparse==0.4.4
    # via
    #   -r requirements.in
    #   django
tomli==2.0.1
    # via
    #   -r requirements.in
    #   black
    #   build
    #   pyproject-hooks
typing-extensions==4.6.2
    # via
    #   -r requirements.in
    #   asgiref
    #   black
virtualenv==20.23.0
    # via
    #   -r requirements.in
    #   pre-commit
wheel==0.40.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
